More refactoring of composite sample wip
