+++ Patterns/Composite/App/compose/Models/Generic/CompositeNode.cs
@@ -12,5 +12,6 @@
 
     public override bool IsLeaf { get; protected set; } = false;
 
+    public override bool LockRoot { get; protected set; } = false;
   }
 }
